model: implement Program interface
The only real challenge here is making sure the underlying collections
are all immutable ("frozen").

Collections.unmodifiableMap makes it impossible to modify the map, but
not the underlying keys or values; to freeze the values along with the
map, we
- construct an intermediate map,
- replace all the values with immutable variations, and finally
- freeze the map.

The replaceAll(Bifunction<K,V,V>) method is concise here; there is no
stream-based equivalent for maps without streaming over .entrySet(),
probably in combination with .forEach() applying .setValue(), and that
seems likely to run into ConcurrentModificationExceptions. A for-loop
would do, but concision + clarity rules.