Program: add and test depth
Note the following:
- define depth(R) where R ∈ ignore ≡ 0 (where ignore is the set of rules
  whose depth we are recursively computing so far--this cuts the cycle
  between rules in order to avoid infinite loops, and is deterministic)
- define depth(R) where R ∈ undefined ≡ 1
- define depth(R) where R is a Fact ≡ 0
- define depth(R) for all other R as 1 + the maximum value over the
  depth of all its subrules (ignoring R)

Memoizing won't help the repeated recursion here! While it is true that
the depth function will recursively compute intermediate depths, they
are not "true"/"top-level" depths; that is, an intermediate (private)
depth of R may not be the same as depth(R) given by the public method,
as it depends on the ignore set at the time (holding the undefined set
constant).

One could memoize private depth but would need to use all 3 arguments as
keys; benchmarks are needed to prove that this is even advantageous (I
believe it is unlikely that private depth is repeatedly called on the
same rule with the same ignore set).

One could memoize the public function, so that repeated calls with the
same name avoid repeated computation. For now, we leave it to the
consumer to perform this optimization---don't ask for expensive work
more than once if you can avoid it.