fix LEQ matcher
We cannot use a "path-length" to find the parent Term context, because
self-recursive rules cause the paths to vary in length.

Further, we cannot actually use the parent term context; that context
contains the tree containing the operator and both operands!

The "best" super-type is (unfortunately) ParseTree---we cannot even be
more specific, because some operators have operands of different types!
That is, we cannot type-parameterize the implementations of
InvocationMatcher, for we would not be able to put them all in a
List<T>. They would have to be List<ParseTree> anyways.

So, we throw away some information, and we accept this. Fortunately, for
our purposes, I don't think it will affect much.