add all Binary700 operators
This necessitated changing from Parser::compileParseTreePattern to
ParseTreePatternMatcher::compile so that we could use
ParseTreePatternMatcher::setDelimiters---we need to change the escape
character, because it cannot match in the pattern!

In particular, we need to match operators containing \ (java: "\\"), and
that is the default escape character. A sample
ParseTreePatternMatcher::split with different patterns reveals that (as
far as I am aware) you cannot have the escape character in the pattern
(doubling does not work).

I thus chose (somewhat at random) `, which is not used by any Prolog
standard operators, and nicely resembles Lisps quasi-quote.