convert "bodies" to TermContext
These right-hand-sides are now "term" and not "termlist."

I investigated integrating ParserUtils::invocationPattern and the various
things that use it into the new InvocationMatcher design pattern, but it
would require some changes. The name of a rule found by
ParserUtils::invocationPattern depends on the match, so I would have to
change InvocationMatcher::name to take the match object as well.

Given that it otherwise works fine, I see no reason to change things
(for now).