reintegrate rule/rule0 handling
We use the ParseTreeMatch to pull out names, arguments, & body, as it
gives us a bit of cleaner interface. Technically we can also get the
body from ctx.term(1), and we could pull the names and arguments out of
ctx.term(0) if we were so inclined. But it's a deep tree, so we don't
like that idea.

The only other change is adding ',' to the list of undefined rules in
the test. Now that we properly detect its usage, it shows up in our
tests, and we need to account for it.